// lib/plant_tracker/model/areaOut.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Book Store API
 * Book Store API for renting books and notifying available/returned books in a store
 * OpenAPI spec version: 1.0.0
 */
import type { LocationOut } from "./locationOut";
import type { UserSchema } from "./userSchema";

export interface AreaOut {
  id: string;
  location: LocationOut;
  name: string;
  user: UserSchema;
}